+++ src/main/php/com/github/mustache/FilesIn.class.php
@@ -2,39 +2,33 @@
 
 use io\Folder;
 use io\File;
-use io\FileUtil;
 
 /**
- * File-based template loading loads templates from the file system.
+ * File-based template loading loads templates from a given folder
  */
-class FilesIn extends \lang\Object implements TemplateLoader {
-  protected $base;
+class FilesIn extends FileBasedTemplateLoader {
 
   /**
    * Creates a new file-based template loader
    *
    * @param var $base The base folder, either an io.Folder or a string
+   * @param string[] $extensions File extensions to check, including leading "."
    */
-  public function __construct($arg) {
-    if ($arg instanceof Folder) {
-      $this->base= $arg;
-    } else {
-      $this->base= new Folder($arg);
-    }
+  public function __construct($arg, $extensions= array('.mustache')) {
+    parent::__construct(
+      $arg instanceof Folder ? $arg : new Folder($arg),
+      $extensions
+    );
   }
 
   /**
-   * Load a template by a given name
+   * Returns an inputstream for a given name, or NULL
    *
-   * @param  string $name The template name without file extension
+   * @param  string $name
    * @return io.streams.InputStream
-   * @throws com.github.mustache.TemplateNotFoundException
    */
-  public function load($name) {
-    $template= new File($this->base, $name.'.mustache');
-    if (!$template->exists()) {
-      throw new TemplateNotFoundException('Cannot find template '.$name.'.mustache in '.$this->base->getURI());
-    }
-    return $template->getInputStream();
+  protected function inputStreamFor($name) {
+    $f= new File($this->base, $name);
+    return $f->exists() ? $f->getInputStream() : null;
   }
 }