+++ src/main/php/com/github/mustache/ResourcesIn.class.php
@@ -1,13 +1,26 @@
 <?php namespace com\github\mustache;
 
 use lang\IClassLoader;
+use lang\ClassLoader;
 use lang\ElementNotFoundException;
 
 /**
  * Classloader template loading loads templates from the class path.
  */
-class ResourcesIn extends \lang\Object implements TemplateLoader {
-  protected $loader;
+class ResourcesIn extends FileBasedTemplateLoader {
+
+  /**
+   * Creates a new class loader based template loader
+   *
+   * @param var $base The delegate, either an IClassLoader or a string
+   * @param string[] $extensions File extensions to check, including leading "."
+   */
+  public function __construct($arg, $extensions= array('.mustache')) {
+    parent::__construct(
+      $arg instanceof IClassLoader ? $arg : ClassLoader::registerPath($arg),
+      $extensions
+    );
+  }
 
   /**
    * Creates a new class loader based template loader instance